Increased the buffer size in uWSGI to address wagtail preview issue
uWSGI allocates a default buffer size of 4096 for request headers. When using the preview functionality the header is much larger than that buffer size so we raise a 502 error due to uWSGI failing to process the request. Increasing the buffer size allows for larger header values, preventing the failure on draft previews.