Add git ref to Github action 'uses' specifier (#4696)
In the "scanning.yml" workflow, add '@ref' notation to the 'uses'
properties. This is required by Github and we are getting errors without
them. I'm adding '@master' to the zaproxy actions in order to get the
latest vulnerability definitions without having to update this file
whenever they do a new release.