remove istanbul plugin and node-sass
node-sass is now deprecated, and the pure-JS sass package is preferred.
Getting rid of the former in favor of the latter removes a native code
dependency (hooray!) which has often been a major headache for us in the
past.

I also realized that I don't think we are using babel-plugin-istanbul
for anything so it can be removed.

pr #4965