uprade nodejs to v14
This upgrades nodejs to v14.18.2. This This isn't fully updated (current
version is 17.x.x) but this will get us to the point where we can
upgrade yarn to v3 and run some other upgrades that will then, I think,
let us upgrade to a more recent nodejs version and so on. We're climbing
the ladder here :)

part of #5129
pr #5130